Move IoU functions to metrics.py
